Move .inboundGateway() and .outboundGateway() to 'msgs/gateways'
The inbound and outbound gateway are the only components that need
when.js. By moving them into a submodule we can remove the required
dependency on when.js.

For Node, when.js is now an optionalDependency. As Bower doesn't support
optional dependenices, it remains a regular dependency in component.json.